list-accounts: improve performance (#54)
If smwingsd is not running, the call to getKey function is converted
to a call to the 'db' command. Even on a database with a small number
of records, such implementation has a huge impact on performances.

The fix reads calls the 'db' command just once and store the results
in RAM. On a database with only 30 records, the script is now 10x
faster.